auth_api/python/requirements.txt
Flask==2.2.2
mysql-connector-python==8.0.32
PyJWT==2.6.0